Move data retrieval to StreamReader class
