+++ lib/FastImageSize.php
@@ -13,14 +13,11 @@ namespace FastImageSize;
 
 class FastImageSize
 {
-	/** @var bool Flag whether allow_url_fopen is enabled */
-	protected $isFopenEnabled = false;
+	/** @var StreamReader */
+	protected $streamReader;
 
 	/** @var array Size info that is returned */
 	protected $size = array();
-
-	/** @var string Data retrieved from remote */
-	protected $data = '';
 
 	/** @var array List of supported image types and associated image types */
 	protected $supportedTypes = array(
@@ -79,13 +76,12 @@ class FastImageSize
 	 */
 	public function __construct()
 	{
-		$iniGet = new \bantu\IniGetWrapper\IniGetWrapper();
-		$this->isFopenEnabled = $iniGet->getBool('allow_url_fopen');
+		$this->streamReader = new StreamReader();
 
 		foreach ($this->supportedTypes as $imageType => $extension)
 		{
 			$className = '\FastImageSize\Type\Type' . mb_convert_case(mb_strtolower($imageType), MB_CASE_TITLE);
-			$this->type[$imageType] = new $className($this);
+			$this->type[$imageType] = new $className($this, $this->streamReader);
 
 			// Create class map
 			foreach ($extension as $ext)
@@ -141,7 +137,7 @@ class FastImageSize
 	protected function getImagesizeUnknownType($filename)
 	{
 		// Grab the maximum amount of bytes we might need
-		$data = $this->getImage($filename, 0, Type\TypeJpeg::JPEG_MAX_HEADER_SIZE, false);
+		$data = $this->streamReader->getImage($filename, 0, Type\TypeJpeg::JPEG_MAX_HEADER_SIZE, false);
 
 		if ($data !== false)
 		{
@@ -177,7 +173,7 @@ class FastImageSize
 	protected function resetValues()
 	{
 		$this->size = array();
-		$this->data = '';
+		$this->streamReader->resetData();
 	}
 
 	/**
@@ -201,34 +197,6 @@ class FastImageSize
 	}
 
 	/**
-	 * Get image from specified path/source
-	 *
-	 * @param string $filename Path to image
-	 * @param int $offset Offset at which reading of the image should start
-	 * @param int $length Maximum length that should be read
-	 * @param bool $forceLength True if the length needs to be the specified
-	 *			length, false if not. Default: true
-	 *
-	 * @return false|string Image data or false if result was empty
-	 */
-	public function getImage($filename, $offset, $length, $forceLength = true)
-	{
-		if (empty($this->data))
-		{
-			$this->getImageData($filename, $offset, $length);
-		}
-
-		// Force length to expected one. Return false if data length
-		// is smaller than expected length
-		if ($forceLength === true)
-		{
-			return (strlen($this->data) < $length) ? false : substr($this->data, $offset, $length) ;
-		}
-
-		return empty($this->data) ? false : $this->data;
-	}
-
-	/**
 	 * Get return data
 	 *
 	 * @return array|bool Size array if dimensions could be found, false if not
@@ -236,68 +204,5 @@ class FastImageSize
 	protected function getReturnData()
 	{
 		return sizeof($this->size) > 1 ? $this->size : false;
-	}
-
-	/**
-	 * Get image data for specified filename with offset and length
-	 *
-	 * @param string $filename Path to image
-	 * @param int $offset Offset at which reading of the image should start
-	 * @param int $length Maximum length that should be read
-	 */
-	protected function getImageData($filename, $offset, $length)
-	{
-		// Check if we don't have a valid scheme according to RFC 3986 and
-		// try to use file_get_contents in that case
-		if (preg_match('#^([a-z][a-z0-9+\-.]+://)#i', $filename))
-		{
-			try
-			{
-				$body = $this->getSeekableImageData($filename, $offset);
-
-				while (!$body->eof())
-				{
-					$readLength = min($length - strlen($this->data), 8192);
-					$this->data .= $body->read($readLength);
-					if ($readLength < 8192 || strlen($this->data == $readLength))
-					{
-						break;
-					}
-				}
-			}
-			catch (\GuzzleHttp\Exception\RequestException $exception)
-			{
-				// Silently fail in case of issues during guzzle request
-			}
-		}
-
-		if (empty($this->data) && $this->isFopenEnabled)
-		{
-			$this->data = @file_get_contents($filename, null, null, $offset, $length);
-		}
-	}
-
-	/**
-	 * Get seekable image data in form of Guzzle stream interface
-	 *
-	 * @param string $filename Filename / URL to get
-	 * @param int $offset Offset for response body
-	 * @return \GuzzleHttp\Stream\StreamInterface|null Stream interface of
-	 *		requested image or null if it could not be retrieved
-	 */
-	public function getSeekableImageData($filename, $offset)
-	{
-		$guzzleClient = new \GuzzleHttp\Client();
-		// Set stream to true to not read full file data during request
-		$response = $guzzleClient->get($filename, ['stream' => true]);
-
-		$body = $response->getBody();
-
-		if ($offset > 0 && !$body->eof())
-		{
-			$body->seek($offset);
-		}
-
-		return $body;
 	}
 }
