+++ lib/Type/TypeBase.php
@@ -12,19 +12,25 @@
 namespace FastImageSize\Type;
 
 use \FastImageSize\FastImageSize;
+use FastImageSize\StreamReader;
 
 abstract class TypeBase implements TypeInterface
 {
 	/** @var FastImageSize */
 	protected $fastImageSize;
 
+	/** @var StreamReader */
+	protected $streamReader;
+
 	/**
 	 * Base constructor for image types
 	 *
 	 * @param FastImageSize $fastImageSize
+	 * @param StreamReader $streamReader
 	 */
-	public function __construct(FastImageSize $fastImageSize)
+	public function __construct(FastImageSize $fastImageSize, StreamReader $streamReader)
 	{
 		$this->fastImageSize = $fastImageSize;
+		$this->streamReader = $streamReader;
 	}
 }
