+++ lib/Type/TypePng.php
@@ -26,7 +26,7 @@ class TypePng extends TypeBase
 	{
 		// Retrieve image data including the header, the IHDR tag, and the
 		// following 2 chunks for the image width and height
-		$data = $this->fastImageSize->getImage($filename, 0, self::PNG_IHDR_OFFSET + 3 * self::LONG_SIZE);
+		$data = $this->streamReader->getImage($filename, 0, self::PNG_IHDR_OFFSET + 3 * self::LONG_SIZE);
 
 		// Check if header fits expected format specified by RFC 2083
 		if (substr($data, 0, self::PNG_IHDR_OFFSET - self::LONG_SIZE) !== self::PNG_HEADER || substr($data, self::PNG_IHDR_OFFSET, self::LONG_SIZE) !== 'IHDR')
