+++ lib/Type/TypePsd.php
@@ -27,7 +27,7 @@ class TypePsd extends TypeBase
 	 */
 	public function getSize($filename)
 	{
-		$data = $this->fastImageSize->getImage($filename, 0, self::PSD_HEADER_SIZE);
+		$data = $this->streamReader->getImage($filename, 0, self::PSD_HEADER_SIZE);
 
 		if ($data === false)
 		{
