+++ lib/Type/TypeTif.php
@@ -50,7 +50,7 @@ class TypeTif extends TypeBase
 	public function getSize($filename)
 	{
 		// Do not force length of header
-		$data = $this->fastImageSize->getImage($filename, 0, self::TIF_HEADER_SIZE, false);
+		$data = $this->streamReader->getImage($filename, 0, self::TIF_HEADER_SIZE, false);
 
 		$this->size = array();
